+++ STM32/InvenSense/spi.h
@@ -19,7 +19,7 @@ void MPU_NSS_High(void);
 void MPU_WriteByte(uint8_t address, uint8_t data);
 uint8_t MPU_ReadByte(uint8_t address);
 void MPU_Write(uint8_t address, uint8_t *data, uint8_t size);
-void MPU_Read(uint8_t address, uint8_t *data, uint8_t size);
+void MPU_Read(uint8_t address, uint8_t *data, uint16_t size);
 
 void SPI2_Init(void);
